script/train.py
#!/home/dawnat9/miniconda3/envs/llm/bin/python
"""
TOMAS-LLM 训练脚本
"""

import argparse
import logging
from pathlib import Path

import yaml


def load_config(config_path: str):
    """加载配置文件"""
    with open(config_path, 'r', encoding='utf-8') as f:
        config = yaml.safe_load(f)
    return config


def main():
    """训练主函数"""
    parser = argparse.ArgumentParser(description="TOMAS-LLM Training Script")
    parser.add_argument(
        "--config",
        type=str,
        default="configs/default.yaml",
        help="Path to configuration file"
    )
    parser.add_argument(
        "--resume",
        type=str,
        default=None,
        help="Path to checkpoint to resume training"
    )
    
    args = parser.parse_args()
    
    # 加载配置
    config = load_config(args.config)
    
    # 配置日志
    logging.basicConfig(
        level=logging.INFO,
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
    )
    logger = logging.getLogger(__name__)
    
    logger.info("=" * 60)
    logger.info("TOMAS-LLM Training Starting...")
    logger.info("=" * 60)
    
    # TODO: 实现训练流程
    # 1. 加载数据集
    # 2. 初始化模型（Tool Encoder, Resource MLP, Temporal Encoder, Qwen Backbone, Output Heads）
    # 3. 配置优化器和学习率调度
    # 4. 训练循环
    # 5. 验证和保存 checkpoint
    
    logger.info("Training configuration:")
    logger.info(f"  Batch size: {config['training']['batch_size']}")
    logger.info(f"  Epochs: {config['training']['num_epochs']}")
    logger.info(f"  Learning rate: {config['training']['learning_rate']}")
    logger.info(f"  Strategy: {config['training']['strategy']}")
    
    logger.warning("Training pipeline not yet implemented - placeholder only")
    
    logger.info("=" * 60)
    logger.info("Training Finished")
    logger.info("=" * 60)


if __name__ == "__main__":
    main()
