requirements.txt
# TOMAS-LLM Dependencies

# Core ML frameworks
torch>=2.0.0
transformers>=4.36.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Training utilities
accelerate>=0.25.0
peft>=0.7.0  # For LoRA

# Evaluation
scikit-learn>=1.3.0

# Logging and utilities
tqdm>=4.65.0
